Plenty of compiler warning fixes
* Mainly int-uint static casts
* Also some nullptr changes

+++ xvisourcesink.h
@@ -27,7 +27,7 @@ private:
     uint32_t bitsperpixel;
     QString dataformat;
     int rows,cols;
-    int nFrames;
+    ulong nFrames;
     int currPos;
 };
